FRONT: apply requested changes
MINOR
